Removed BaseTest and corresponding code
