+++ src/test/java/uk/gov/hmcts/reform/em/orchestrator/endpoint/CcdCloneBundleControllerTest.java
@@ -1,15 +1,36 @@
 package uk.gov.hmcts.reform.em.orchestrator.endpoint;
 
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
+import org.springframework.security.oauth2.core.oidc.OidcIdToken;
+import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser;
+import org.springframework.security.oauth2.core.oidc.user.OidcUser;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestFilter;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
 import uk.gov.hmcts.reform.em.orchestrator.Application;
 import uk.gov.hmcts.reform.em.orchestrator.service.caseupdater.CcdBundleCloningService;
+import uk.gov.hmcts.reform.em.orchestrator.service.caseupdater.DefaultUpdateCaller;
 
 import javax.servlet.http.HttpServletRequest;
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
@@ -17,7 +38,50 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = {Application.class})
-public class CcdCloneBundleControllerTest extends BaseTest {
+public class CcdCloneBundleControllerTest {
+
+    @MockBean
+    private DefaultUpdateCaller defaultUpdateCaller;
+
+    @MockBean
+    private JwtDecoder jwtDecoder;
+
+    @Autowired
+    private WebApplicationContext context;
+
+    private static final String ID_TOKEN = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9"
+            + ".eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsIm"
+            + "p0aSI6ImQzNWRmMTRkLTA5ZjYtNDhmZi04YTkzLTdjNmYwMzM5MzE1OSIsImlhdCI6MTU0M"
+            + "Tk3MTU4MywiZXhwIjoxNTQxOTc1MTgzfQ.QaQOarmV8xEUYV7yvWzX3cUE_4W1luMcWCwpr"
+            + "oqqUrg";
+
+    private MockMvc restLogoutMockMvc;
+
+    @Before
+    public void setupMocks() {
+
+        Map<String, Object> claims = new HashMap<>();
+        claims.put("groups", "ROLE_USER");
+        claims.put("sub", 123);
+
+        OidcIdToken idToken = new OidcIdToken(ID_TOKEN, Instant.now(),
+                Instant.now().plusSeconds(60), claims);
+
+        SecurityContextHolder.getContext().setAuthentication(authenticationToken(idToken));
+
+        SecurityContextHolderAwareRequestFilter authInjector = new SecurityContextHolderAwareRequestFilter();
+        this.restLogoutMockMvc = MockMvcBuilders.webAppContextSetup(this.context).build();
+    }
+
+    private OAuth2AuthenticationToken authenticationToken(OidcIdToken idToken) {
+
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(new SimpleGrantedAuthority("USER"));
+        OidcUser user = new DefaultOidcUser(authorities, idToken);
+
+        return new OAuth2AuthenticationToken(user, authorities, "oidc");
+
+    }
 
     @Test
     public void shouldCallCcdCallbackHandlerService() throws Exception {
