+++ src/test/java/uk/gov/hmcts/reform/em/orchestrator/endpoint/StitchingCompleteCallbackControllerTest.java
@@ -6,11 +6,24 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
+import org.springframework.security.oauth2.core.oidc.OidcIdToken;
+import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser;
+import org.springframework.security.oauth2.core.oidc.user.OidcUser;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestFilter;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
 import uk.gov.hmcts.reform.em.orchestrator.Application;
 import uk.gov.hmcts.reform.em.orchestrator.service.notification.NotificationService;
 import uk.gov.hmcts.reform.em.orchestrator.service.orchestratorcallbackhandler.CallbackException;
@@ -20,7 +33,8 @@ import uk.gov.hmcts.reform.em.orchestrator.stitching.dto.StitchingBundleDTO;
 import uk.gov.hmcts.reform.em.orchestrator.stitching.dto.TaskState;
 
 import java.io.IOException;
-import java.util.UUID;
+import java.time.Instant;
+import java.util.*;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
@@ -31,7 +45,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = {Application.class})
 @AutoConfigureMockMvc
-public class StitchingCompleteCallbackControllerTest extends BaseTest {
+public class StitchingCompleteCallbackControllerTest {
 
     @MockBean
     private StitchingCompleteCallbackService stitchingCompleteCallbackService;
@@ -39,10 +53,37 @@ public class StitchingCompleteCallbackControllerTest extends BaseTest {
     @MockBean
     private NotificationService notificationService;
 
+    @MockBean
+    private JwtDecoder jwtDecoder;
+
+    @Autowired
+    private WebApplicationContext context;
+
+    private static final String ID_TOKEN = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9"
+            + ".eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsIm"
+            + "p0aSI6ImQzNWRmMTRkLTA5ZjYtNDhmZi04YTkzLTdjNmYwMzM5MzE1OSIsImlhdCI6MTU0M"
+            + "Tk3MTU4MywiZXhwIjoxNTQxOTc1MTgzfQ.QaQOarmV8xEUYV7yvWzX3cUE_4W1luMcWCwpr"
+            + "oqqUrg";
+
+    private MockMvc restLogoutMockMvc;
+
     private String requestBody;
 
     @Before
     public void setUp() throws IOException {
+
+        Map<String, Object> claims = new HashMap<>();
+        claims.put("groups", "ROLE_USER");
+        claims.put("sub", 123);
+
+        OidcIdToken idToken = new OidcIdToken(ID_TOKEN, Instant.now(),
+                Instant.now().plusSeconds(60), claims);
+
+        SecurityContextHolder.getContext().setAuthentication(authenticationToken(idToken));
+
+        SecurityContextHolderAwareRequestFilter authInjector = new SecurityContextHolderAwareRequestFilter();
+        this.restLogoutMockMvc = MockMvcBuilders.webAppContextSetup(this.context).build();
+
         DocumentTaskDTO documentTaskDTO = new DocumentTaskDTO();
         StitchingBundleDTO stitchingBundleDTO = new StitchingBundleDTO();
         stitchingBundleDTO.setEnableEmailNotification(true);
@@ -92,6 +133,16 @@ public class StitchingCompleteCallbackControllerTest extends BaseTest {
                 .andExpect(status().is(456))
                 .andExpect(jsonPath("$.message", Matchers.is("error message")))
                 .andExpect(jsonPath("$.httpResponseBody", Matchers.is("error")));
+
+    }
+
+    private OAuth2AuthenticationToken authenticationToken(OidcIdToken idToken) {
+
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(new SimpleGrantedAuthority("USER"));
+        OidcUser user = new DefaultOidcUser(authorities, idToken);
+
+        return new OAuth2AuthenticationToken(user, authorities, "oidc");
 
     }
 }
